+++ PhpSerializerNET.Test/Deserialize/IntegerDeserialization.cs
@@ -41,5 +41,17 @@ namespace PhpSerializerNET.Test {
 				PhpSerialization.Deserialize<int>("i:-2147483648;")
 			);
 		}
+
+		[TestMethod]
+		public void ExplictCastFormatException() {
+			var ex = Assert.ThrowsException<PhpSerializerNET.DeserializationException>( () => 
+				PhpSerialization.Deserialize<int>(
+					"s:3:\"1b1\";"
+				)
+			);
+			Assert.IsInstanceOfType(ex.InnerException, typeof(System.FormatException));
+			Assert.AreEqual("Exception encountered while trying to assign '1b1' to type Int32. See inner exception for details.", ex.Message);
+		}
+
 	}
 }