+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Object / struct creation: Improved performance.
 - General: Reduced amount of memory allocated while deserializing.
 - Fixed exception message for non-integer keys in lists.
+- Fixed exception message for failed field / property assignments / binding.
 
 # 0.8.0:
 - Improved performance of the validation step of deserialization.